agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "deepagents>=0.4.5",
    "fastapi>=0.135.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "letta-client>=1.7.8",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "python-toon>=0.1.3",
    "uvicorn[standard]>=0.41.0",
]
